Uploading Abstract (changes as per comments-done)
Added Animal bites to the moi categorization code
Merged two csv files to have the ISS scores and the latest follow up data.
Tried changing the datatypes.. But it didn't work. The code I used was introducing NAs. 
Tried the complete case analysis too.. I was trying to convert 999 to NA of only particular columns.. and then tried the complete case analysis code. But even that was introducing NAs. 
I haven't included these codes since this was supposed to be the final submission.